+++ NEWS.md
@@ -72,6 +72,7 @@ New library functions
 ---------------------
 
 * `Iterators.flatmap` was added ([#44792]).
+* `@cglobal`, macro version of `cglobal` to access global variables ([#????]).
 * New helper `Splat(f)` which acts like `x -> f(x...)`, with pretty printing for
   inspecting which function `f` was originally wrapped. ([#42717])
 * New `pkgversion(m::Module)` function to get the version of the package that loaded